Move SingleItemHandler to lib
